feat: improved inline component experience
