+++ packages/kit-headless/src/components/accordion/accordion-root.tsx
@@ -1,4 +1,6 @@
 import {
+  PropsOf,
+  QRL,
   Signal,
   Slot,
   component$,
@@ -7,8 +9,41 @@ import {
   useSignal,
   useTask$,
 } from '@builder.io/qwik';
-import { AccordionRootProps } from './accordion-inline';
 import { accordionContextId } from './accordion-context';
+
+export type AccordionRootProps = PropsOf<'div'> & {
+  /** If true, multiple items can be open at the same time. */
+  multiple?: boolean;
+
+  /**
+   * @deprecated Use the multiple prop instead.
+   */
+  behavior?: 'single' | 'multi';
+
+  /** The reactive value controlling which item is open. */
+  'bind:value'?: Signal<string | null>;
+
+  /** The initial value of the currently open item. */
+  value?: string;
+
+  /** The initial index of the currently open item. */
+  initialIndex?: number;
+
+  /** A QRL that is called when the selected item changes. */
+  onChange$?: QRL<(value: string) => void>;
+
+  /** A map of the item indexes and their disabled state. */
+  itemsMap?: Map<number, boolean>;
+
+  /** If true, the accordion is disabled. */
+  disabled?: boolean;
+
+  /** If true, the accordion is collapsible. */
+  collapsible?: boolean;
+
+  /** If true, the accordion is animated. */
+  animated?: boolean;
+};
 
 export const HAccordionRootImpl = component$((props: AccordionRootProps) => {
   const {
